requirements.txt
aiohttp==3.9.5
aiohttp-retry==2.8.3
aiosignal==1.3.1
amqp==5.2.0
annotated-types==0.6.0
ansible==8.7.0
ansible-core==2.15.11
ansible-vault==2.1.0
antlr4-python3-runtime==4.9.3
anyio==4.3.0
appdirs==1.4.4
async-timeout==4.0.3
asynch==0.2.3
asyncssh==2.14.2
atpublic==4.1.0
attrs==23.2.0
billiard==4.2.0
celery==5.4.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
ciso8601==2.3.1
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
clickhouse-connect==0.7.8
clickhouse-driver==0.2.7
colorama==0.4.6
configobj==5.0.8
configparser==7.0.0
cryptography==42.0.5
dictdiffer==0.9.0
diskcache==5.6.3
distro==1.9.0
dpath==2.1.6
dulwich==0.21.7
dvc==3.50.0
dvc-data==3.15.1
dvc-http==2.32.0
dvc-objects==5.1.0
dvc-render==1.0.2
dvc-studio-client==0.20.0
dvc-task==0.4.0
entrypoints==0.4
exceptiongroup==1.2.0
fastapi==0.110.1
filelock==3.13.4
flatten-dict==0.4.2
flufl.lock==7.1.1
frozenlist==1.4.1
fsspec==2024.3.1
funcy==2.0
gitdb==4.0.11
GitPython==3.1.43
grandalf==0.8
gto==1.7.1
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
hydra-core==1.3.2
idna==3.7
importlib-resources==5.0.7
iniconfig==2.0.0
iterative-telemetry==0.0.8
Jinja2==3.1.3
joblib==1.4.0
kafka-python==2.0.2
kombu==5.3.7
leb128==1.0.7
lz4==4.3.3
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
multidict==6.0.5
networkx==3.2.1
nltk==3.8.1
numpy==1.26.4
omegaconf==2.3.0
orjson==3.10.1
packaging==24.0
pandas==2.2.2
path==16.14.0
pathspec==0.12.1
pip==21.2.4
platformdirs==3.11.0
pluggy==1.4.0
prompt-toolkit==3.0.43
psutil==5.9.8
pycparser==2.22
pydantic==2.7.0
pydantic_core==2.18.1
pydot==2.0.0
pygit2==1.14.1
Pygments==2.17.2
pygtrie==2.5.0
pyparsing==3.1.2
pytest==8.1.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
regex==2023.12.25
requests==2.31.0
resolvelib==1.0.1
rich==13.7.1
ruamel.yaml==0.18.6
ruamel.yaml.clib==0.2.8
scikit-learn==1.4.2
scipy==1.13.0
scmrepo==3.3.1
semver==3.0.2
setuptools==58.0.4
shellingham==1.5.4
shortuuid==1.0.13
shtab==1.7.1
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
sqltrie==0.11.0
starlette==0.37.2
tabulate==0.9.0
threadpoolctl==3.4.0
tomli==2.0.1
tomlkit==0.12.4
tqdm==4.66.2
typer==0.12.3
typing_extensions==4.11.0
tzdata==2024.1
tzlocal==5.2
urllib3==2.2.1
uvicorn==0.29.0
vine==5.1.0
voluptuous==0.14.2
wcwidth==0.2.13
yarl==1.9.4
zc.lockfile==3.0.post1
zstandard==0.22.0
zstd==1.5.5.1
